do code format for all except reader-gtfs #1451

+++ core/src/main/java/com/graphhopper/apache/commons/collections/IntDoubleBinaryHeap.java
@@ -82,7 +82,7 @@ public class IntDoubleBinaryHeap implements BinHeapWrapper<Number, Integer> {
 
     public void insert_(double key, int element) {
         if (isFull()) {
-            ensureCapacity((int) (elements.length * GROW_FACTOR));
+            ensureCapacity(elements.length * GROW_FACTOR);
         }
 
         size++;
